Add phoogle_folder_id IAM folder ID to KMS for 0.12
Without this patch the integration tests for IAM create resources at the
organization level where it does not have sufficient permissions.  This
patch is a necessary step to plumb the IAM Folder ID through the
Concourse Pipeline to the integration tests so that resources may be
created in the IAM folder fixture.